PS-7: restructure + refractor
